requirements.txt
python-dotenv
openai == 0.28.1
langchain == 0.0.312
pandas